Rename datetime rebase module into datetime utils module
