+++ src/main/cpp/tests/datetime_rebase.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include <datetime_rebase.hpp>
+#include <datetime_utils.hpp>
 //
 
 #include <cudf_test/base_fixture.hpp>
