+++ src/main/java/com/nvidia/spark/rapids/jni/DateTimeRebase.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, NVIDIA CORPORATION.
+ * Copyright (c) 2023-2024, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,41 +19,20 @@ package com.nvidia.spark.rapids.jni;
 import ai.rapids.cudf.*;
 
 /**
- * Utility class for converting between column major and row major data
+ * @deprecated Use {@link DateTimeUtils} instead.
+ *
+ * This will be removed after the plugin picks up DateTimeUtils class.
  */
 public class DateTimeRebase {
   static {
     NativeDepsLoader.loadNativeDeps();
   }
 
-  /**
-   * Convert the given timestamps as a number of days or microseconds since the epoch instant
-   * 1970-01-01T00:00:00Z to a local date-time in Proleptic Gregorian calendar, reinterpreting
-   * the result as in Julian calendar, then compute the number of days or microseconds since the
-   * epoch from that Julian local date-time.
-   * <p>
-   * This is to match with Apache Spark's `localRebaseGregorianToJulianDays` and
-   * `rebaseGregorianToJulianMicros` functions with timezone fixed to UTC.
-   */
   public static ColumnVector rebaseGregorianToJulian(ColumnView input) {
-    return new ColumnVector(rebaseGregorianToJulian(input.getNativeView()));
+    return DateTimeUtils.rebaseGregorianToJulian(input);
   }
 
-  /**
-   * Convert the given timestamps as a number of days or microseconds since the epoch instant
-   * 1970-01-01T00:00:00Z to a local date-time in Julian calendar, reinterpreting the result
-   * as in Proleptic Gregorian calendar, then compute the number of days or microseconds since the
-   * epoch from that Gregorian local date-time.
-   * <p>
-   * This is to match with Apache Spark's `localRebaseJulianToGregorianDays` and
-   * `rebaseJulianToGregorianMicros` functions with timezone fixed to UTC.
-   */
   public static ColumnVector rebaseJulianToGregorian(ColumnView input) {
-    return new ColumnVector(rebaseJulianToGregorian(input.getNativeView()));
+    return DateTimeUtils.rebaseJulianToGregorian(input);
   }
-
-
-  private static native long rebaseGregorianToJulian(long nativeHandle);
-
-  private static native long rebaseJulianToGregorian(long nativeHandle);
 }
